+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG.md
 
+## 0.2.8 (2023-04-27)
+
+- Added `shouldDirty` and `shouldTouch` to `setValue` call in `FormFileInput` component so form state is updated correctly when file is selected.
+- Added `FormTextarea` component which integrates DaisyUI's `Textarea` with react-hook-form.
+
 ## 0.2.7 (2023-04-26)
 
 - Added `FormFileInput` component that adds React Hook Form integration to DaisyUI's `FileInput`.